added view-model

+++ app/src/main/java/com/example/trempel/network/model/ProductModel.kt
@@ -5,7 +5,7 @@ internal data class ProductModel(
     val description: String,
     val id: Int,
     val image: String,
-    val price: Double,
+    val price: Float,
     val rating: Rating,
     val title: String
 )
